Re-order queue so that all tasks progress at about same speed
Currently the queue takes all builds from one task before beginning
with the next. So if one wants to play many builds on the same repo
for multiple tasks, it can take long before results can be compared.
With this CL, build A will be executed for those tasks before moving
to build B, thus solving this issue.