WIP Change bot queue management
This CL changes the use of task.base by moving it from
"current ToT of task repo towards which the queue is
aiming" to "latest finished build".
This way, is the bot is interrupted, it will be
able to continue where it was instead of thinking it
had finished all task in  the queue.

Fixes #13