Fix the issue that `serie.analyse.base` gotten overridden
After done analysis, a `reduceBaseTo` is set to increase the base
if there are more than 50 samples in the current range, to only
consider the latest 50 samples.

However, it was not properly considering the `indexStart`, which
is determined by the existing `serie.analyse.base`. That can
result in `serie.analyse.base` gotten overridden and we were going
back to some old base commit. This happened in `apiAddSample`,
where we reset the base using `reduceBaseTo`.

This commit fixes the issue.