Add an 'Active' button to filter only active benchmarks
This works by checking each benchmark's latest build ID to be the
largest known build ID. If not, the benchmark is considered as
not active. This could cause false negative when we have benchmark
build IDs going backwards; but that should be fine given it's rare
and should be temporary.